todo funciona TODO
incluso el timer guiño

+++ firmware/pwm_adc_gpio/Core/Inc/command_decoder.h
@@ -12,6 +12,6 @@
 #include "device.h"
 
 int decode_pc_command(device *dev, uint8_t pc_command_1, uint8_t pc_command_2);
-int decode_red_command(device *dev, int red_command_1, int red_command_2);
+int decode_red_command(device *dev, uint8_t red_command_1, uint8_t red_command_2);
 
 #endif /* COMMAND_DECODER_H_ */
